+++ src/app/local-project/index.ts
@@ -1,5 +1,6 @@
 export * from "./context/index.js";
 export * from "./inspection/index.js";
+export * from "./maintenance/index.js";
 export * from "./observation/index.js";
 export * from "./omission/index.js";
 export * from "./setup/index.js";